chore(useBreakpoints): move hook to typescript
