+++ src/index.ts
@@ -28,6 +28,7 @@ import TableDateInput from "./TableDateInput";
 import TableInput from "./TableInput";
 import Tabs from "./Tabs";
 import Tooltip from "./Tooltip";
+import useBreakpoints from "./hooks/useBreakpoints";
 
 /**
  * Untyped Components
@@ -58,7 +59,6 @@ declare const TimelineEvent;
 declare const Toggle;
 declare const TokenInput;
 declare const TruncatedAccount;
-declare const useBreakpoints;
 declare const useLockBodyScroll;
 declare const formatNumber;
 declare const formatDate;
